requirements.txt
scikit-surprise==1.1.4
numpy<2
scipy<1.12
matplotlib
